migration to Neos 2.0
manually change @override to @context

automated migration details:

[TASK] Import core migration log to composer.json
[TASK] Apply migration TYPO3.FLOW3-201201261636

Adjust to the major changes in FLOW3 1.1 when coming from 1.0.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Fluid-201205031303

Rename form.textbox to form.textfield

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.FLOW3-201205292145

Replace FileTypes use with MediaTypes use.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.FLOW3-201206271128

Replace DataNotSerializeableException with DataNotSerializableException.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.FLOW3-201209201112

Move all code to PSR-0 compatible directory structure, remove
Package.xml,

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Flow-201209251426

Rename FLOW3 to TYPO3 Flow

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Flow-201211151101

Check for globally defined role identifiers in Policy.yaml files

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Flow-201212051340

Change entity resource definitions from using _ to \

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.TypoScript-130516234520

Rename "renderTypoScript" VH to just "render"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.TypoScript-130516235550

Rename "renderTypoScript" VH to just "render"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.TYPO3CR-130523180140

Adjust to removed TYPO3.TYPO3CR:Folder node type by replacing it

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Flow-201310031523

Change TYPO3\Flow\Persistence\Doctrine\DatabaseConnectionException to

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Flow-201405111147

This migration does not actually change any code. It just displays a
warning if a TypeConverters canConvertFrom() is

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Neos-201409071922

Adjust to updated folder name for TypoScript in site packages

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.TYPO3CR-140911160326

Adjust to renamed NodeLabelGeneratorInterface.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Neos-201410010000

Migrates the former HtmlEditor to CodeEditor

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.TYPO3CR-141101082142

Adjust namespaces for TYPO3CR related FlowQuery operations that were
moved from TYPO3.Neos to TYPO3.TYPO3CR

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Neos-20141113115300

Rename setting TYPO3.Neos.modules.<moduleName>.resource to
"privilegeTarget"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Fluid-20141113120800

Rename the "resource" argument of the security.ifAccess ViewHelper to
"privilegeTarget"

[TASK] Apply migration TYPO3.Flow-20141113121400

Adjust "Policy.yaml" to new syntax

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Fluid-20141121091700

Warn about removed ReflectionService dependency from AbstractViewHelper

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Neos-20141218134700

Rename node property type 'date' to DateTime

[TASK] Apply migration TYPO3.Fluid-20150214130800

Add "escapeOutput" property to existing ViewHelpers to ensure
backwards-compatibility

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

[TASK] Apply migration TYPO3.Neos-20150303231600

Migrates ImageVariant to ImageInterface

[TASK] Apply migration TYPO3.TYPO3CR-20150510103823

Migrates to new super type syntax in node type configuration