FIX: Prevent accessing non-existing index
Checking if the published item is in scope of the api by comparing the
index breaks if there are less than 10 items, as the current index is
compared to the tenth element's index. If this is null, the condition
breaks. With this fix, the element count is checked to determine whether
the indices have to be compared or not